Initial conversion of 32-bit integers into 64-bit integers.

+++ server.h
@@ -244,7 +244,7 @@ enum Server_Option {
 #define server_int_option_cached(srvopt)   (_server_int_option_cache[srvopt])
 
 
-extern int32 _server_int_option_cache[]; /* private */
+extern int64_t _server_int_option_cache[]; /* private */
